rebase with master

+++ benchmarks/ts/deno/files.ts
@@ -19,11 +19,11 @@ const createBatchOfFiles = (
 ) => {
 	const startTime = performance.now();
 	const fileNames = [];
-	if (!existsSync(path.join(__dirname, directory).toString())) {
+	if (!existsSync(path.join(__dirname, directory).toString(), { isDirectory: true })) {
 		Deno.mkdirSync(path.join(__dirname, directory));
 	}
 	for (let i = 0; i < numberOfFiles; i++) {
-		const fileName = path.join(__dirname, directory, `${startFileName}-${i + 1}.txt`).toString();
+		const fileName = path.join(__dirname, directory, `${startFileName}-${i + 1}.txt`);
 		createFile(fileName, numberOfParagraphs);
 		fileNames.push(fileName);
 	}
@@ -51,15 +51,48 @@ const readFiles = (fileNames: string[]) => {
 };
 
 const removeFiles = (directory = "tmp") => {
-	Deno.removeSync(path.join(__dirname, directory), { recursive: true });
+	if (existsSync(path.join(__dirname, directory)))
+		Deno.removeSync(path.join(__dirname, directory), { recursive: true });
+};
+
+const performFilesBenchmark = (numberOfFiles: number, numberOfParagraphs: number, numberOfIterations: number) => {
+	const result = [];
+	const startTime = performance.now();
+
+	for (let i = 0; i < numberOfIterations; i++) {
+		const { fileNames, time } = createBatchOfFiles(numberOfFiles, "lorem", numberOfParagraphs);
+		const resultReadFiles = readFiles(fileNames);
+		result.push({
+			fileNames,
+			timeToCreateFiles: time,
+			resultsReading: resultReadFiles.results,
+			timeOfReading: resultReadFiles.time,
+		});
+	}
+	const endTime = performance.now();
+	removeFiles();
+
+	return { result, timeToEnd: endTime - startTime };
 };
 
 (() => {
-	const startTime = performance.now();
-	const fileNames = createBatchOfFiles(10, "lorem", 100);
-	const result = readFiles(fileNames.fileNames);
-	const endTime = performance.now();
-	console.log("time ", endTime - startTime);
-	console.log("results ", result);
-	removeFiles();
+	if (Deno.args.length < 3) {
+		Deno.exit(1);
+	}
+
+	const numberOfIterations = Number(Deno.args.at(0));
+	const numberOfFiles = Number(Deno.args.at(1));
+	const numberOfParagraphs = Number(Deno.args.at(2));
+
+	if (!numberOfFiles || !numberOfParagraphs || !numberOfIterations) {
+		Deno.exit(1);
+	}
+
+	const result = performFilesBenchmark(numberOfFiles, numberOfParagraphs, numberOfIterations);
+	const encoder = new TextEncoder();
+	const encodedResult = encoder.encode(JSON.stringify(result));
+
+	Deno.writeFileSync(`${__dirname}/denoFiles.json`, encodedResult);
+
+	Deno.exit(0);
 })();
